Traffice Intelligence
Traffic intelligence is collecting and analyzing data about traffic flow and conditions with the overarching goal of improving mobility and safety. It's built around efficiency in traffic management, further improvement of route planning, and identification of potential accident-prone areas. This information supports urban planning, informs policies, and integrates with smart city technologies. The ultimate goal is to minimize congestion, enhance road safety, and minimize the environmental impact of transportation.